filter duplicated mesh transactions (tx command)
